==> Code/4_Polygenic1.Rmd <==
---
title: "Common Variant Analysis"
output:
  html_document:
    df_print: paged
author:
    Alexandros Rammos, Cardiff University
    
---

```{r}
library(tidyverse)
library(data.table)
library(fmsb)
library(wesanderson)
mypal <- wes_palette("Darjeeling1")

```

Read in the scores generated in Plink from the sample, manipulating them into 1 line/proband

```{r}
pc1 = read.table ("./../data/EA_Trios.all.score", head=T)
pc2 = read.table ("./../data/SZ_Trios.all.score", head=T)
pc3 = read.table ("./../data/IQ_Trios.all.score", head=T)

pc1_1 = subset(pc1, subset=pc1$MEMBERSHIP == "-1"|pc1$MEMBERSHIP == "-4" |pc1$MEMBERSHIP == "-5")
pc1_2 = subset(pc1, subset=pc1$MEMBERSHIP == "-2")
pc1_3 = subset(pc1, subset=pc1$MEMBERSHIP == "-3")

pc2_1 = subset(pc2, subset=pc2$MEMBERSHIP == "-1"|pc2$MEMBERSHIP == "-4" |pc2$MEMBERSHIP == "-5")
pc2_2 = subset(pc2, subset=pc2$MEMBERSHIP == "-2")
pc2_3 = subset(pc2, subset=pc2$MEMBERSHIP == "-3")

pc3_1 = subset(pc3, subset=pc3$MEMBERSHIP == "-1"|pc3$MEMBERSHIP == "-4" |pc3$MEMBERSHIP == "-5")
pc3_2 = subset(pc3, subset=pc3$MEMBERSHIP == "-2")
pc3_3 = subset(pc3, subset=pc3$MEMBERSHIP == "-3")


names(pc1_1) = c("FID","ID","MEMBERSHIP", "EA_Score_0.001", "EA_Score_0.05",  "EA_Score_0.5", "EA_Score_1")
names(pc2_1) = c("FID","ID","MEMBERSHIP", "SZ_Score_0.001", "SZ_Score_0.05",  "SZ_Score_0.5", "SZ_Score_1")
names(pc3_1) = c("FID","ID","MEMBERSHIP", "IQ_Score_0.001", "IQ_Score_0.05",  "IQ_Score_0.5", "IQ_Score_1")


names(pc1_2) = c("FID","Paternal ID","MEMBERSHIP", "EA_Paternal_0.001", "EA_Paternal_0.05",  "EA_Paternal_0.5", "EA_Paternal_1")
names(pc2_2) = c("FID","Paternal ID","MEMBERSHIP", "SZ_Paternal_0.001", "SZ_Paternal_0.05",  "SZ_Paternal_0.5", "SZ_Paternal_1")
names(pc3_2) = c("FID","Paternal ID","MEMBERSHIP", "IQ_Paternal_0.001", "IQ_Paternal_0.05",  "IQ_Paternal_0.5", "IQ_Paternal_1")

names(pc1_3) = c("FID","Maternal ID","MEMBERSHIP", "EA_Maternal_0.001", "EA_Maternal_0.05",  "EA_Maternal_0.5", "EA_Maternal_1")
names(pc2_3) = c("FID","Maternal ID","MEMBERSHIP", "SZ_Maternal_0.001", "SZ_Maternal_0.05",  "SZ_Maternal_0.5", "SZ_Maternal_1")
names(pc3_3) = c("FID","Maternal ID","MEMBERSHIP", "IQ_Maternal_0.001", "IQ_Maternal_0.05",  "IQ_Maternal_0.5", "IQ_Maternal_1")

ea1 = merge (pc1_1, pc1_2, by="FID")
ea1 = merge (ea1, pc1_3, by="FID")

sz1 =merge (pc2_1, pc2_2, by="FID")
sz1 = merge (sz1, pc2_3, by="FID")

iq1 =merge (pc3_1, pc3_2, by="FID")
iq1 = merge (iq1, pc3_3, by="FID")
rm(pc1, pc1_1, pc1_2, pc1_3, pc2, pc2_1, pc2_2, pc2_3,pc3, pc3_1, pc3_2, pc3_3)
```

Next, compute pTDT and ntPRS scores for each proband

```{r}
ea1$EA_Mean_Parental_Score_0.001 = (ea1$EA_Paternal_0.001 + ea1$EA_Maternal_0.001)/2
ea1$EA_pTDT_0.001 = -(ea1$EA_Mean_Parental_Score_0.001 - ea1$EA_Score_0.001)/(sd(ea1$EA_Mean_Parental_Score_0.001))
ea1$EA_ntPRS_0.001 = (ea1$EA_Paternal_0.001 + ea1$EA_Maternal_0.001) - ea1$EA_Score_0.001

ea1$EA_Mean_Parental_Score_0.05 = (ea1$EA_Paternal_0.05 + ea1$EA_Maternal_0.05)/2
ea1$EA_pTDT_0.05 = -(ea1$EA_Mean_Parental_Score_0.05 - ea1$EA_Score_0.05)/(sd(ea1$EA_Mean_Parental_Score_0.05))
ea1$EA_ntPRS_0.05 = (ea1$EA_Paternal_0.05 + ea1$EA_Maternal_0.05) - ea1$EA_Score_0.05


ea1$EA_Mean_Parental_Score_0.5 = (ea1$EA_Paternal_0.5 + ea1$EA_Maternal_0.5)/2
ea1$EA_pTDT_0.5 = -(ea1$EA_Mean_Parental_Score_0.5 - ea1$EA_Score_0.5)/(sd(ea1$EA_Mean_Parental_Score_0.5))
ea1$EA_ntPRS_0.5 = (ea1$EA_Paternal_0.5 + ea1$EA_Maternal_0.5) - ea1$EA_Score_0.5

sz1$SZ_Mean_Parental_Score_0.001 = (sz1$SZ_Paternal_0.001 + sz1$SZ_Maternal_0.001)/2
sz1$SZ_pTDT_0.001 = -(sz1$SZ_Mean_Parental_Score_0.001 - sz1$SZ_Score_0.001)/(sd(sz1$SZ_Mean_Parental_Score_0.001))
sz1$SZ_ntPRS_0.001 = (sz1$SZ_Paternal_0.001 + sz1$SZ_Maternal_0.001) - sz1$SZ_Score_0.001

sz1$SZ_Mean_Parental_Score_0.05 = (sz1$SZ_Paternal_0.05 + sz1$SZ_Maternal_0.05)/2
sz1$SZ_pTDT_0.05 = -(sz1$SZ_Mean_Parental_Score_0.05 - sz1$SZ_Score_0.05)/(sd(sz1$SZ_Mean_Parental_Score_0.05))
sz1$SZ_ntPRS_0.05 = (sz1$SZ_Paternal_0.05 + sz1$SZ_Maternal_0.05) - sz1$SZ_Score_0.05

sz1$SZ_Mean_Parental_Score_0.5 = (sz1$SZ_Paternal_0.5 + sz1$SZ_Maternal_0.5)/2
sz1$SZ_pTDT_0.5 = -(sz1$SZ_Mean_Parental_Score_0.5 - sz1$SZ_Score_0.5)/(sd(sz1$SZ_Mean_Parental_Score_0.5))
sz1$SZ_ntPRS_0.5 = (sz1$SZ_Paternal_0.5 + sz1$SZ_Maternal_0.5) - sz1$SZ_Score_0.5

iq1$IQ_Mean_Parental_Score_0.001 = (iq1$IQ_Paternal_0.001 + iq1$IQ_Maternal_0.001)/2
iq1$IQ_pTDT_0.001 = -(iq1$IQ_Mean_Parental_Score_0.001 - iq1$IQ_Score_0.001)/(sd(iq1$IQ_Mean_Parental_Score_0.001))
iq1$IQ_ntPRS_0.001 = (iq1$IQ_Paternal_0.001 + iq1$IQ_Maternal_0.001) - iq1$IQ_Score_0.001

iq1$IQ_Mean_Parental_Score_0.05 = (iq1$IQ_Paternal_0.05 + iq1$IQ_Maternal_0.05)/2
iq1$IQ_pTDT_0.05 = -(iq1$IQ_Mean_Parental_Score_0.05 - iq1$IQ_Score_0.05)/(sd(iq1$IQ_Mean_Parental_Score_0.05))
iq1$IQ_ntPRS_0.05 = (iq1$IQ_Paternal_0.05 + iq1$IQ_Maternal_0.05) - iq1$IQ_Score_0.05

iq1$IQ_Mean_Parental_Score_0.5 = (iq1$IQ_Paternal_0.5 + iq1$IQ_Maternal_0.5)/2
iq1$IQ_pTDT_0.5 = -(iq1$IQ_Mean_Parental_Score_0.5 - iq1$IQ_Score_0.5)/(sd(iq1$IQ_Mean_Parental_Score_0.5))
iq1$IQ_ntPRS_0.5 = (iq1$IQ_Paternal_0.5 + iq1$IQ_Maternal_0.5) - iq1$IQ_Score_0.5

scores = merge (ea1, iq1, by="ID")
scores = merge (scores, sz1, by="ID")
scores = scores [,-c(3,9,15,29,30,35,36,41,42,56,57,62,63,68,69)] 
rm(ea1, iq1, sz1)
```
Standardize the polygenic scores 

```{r}
scores <- scores %>% 
  mutate_at(c("EA_Score_0.001","EA_Score_0.05","EA_Score_0.5","EA_ntPRS_0.001","EA_ntPRS_0.05","EA_ntPRS_0.5",
              "SZ_Score_0.001","SZ_Score_0.05","SZ_Score_0.5","SZ_ntPRS_0.001","SZ_ntPRS_0.05","SZ_ntPRS_0.5",
              "IQ_Score_0.001","IQ_Score_0.05","IQ_Score_0.5","IQ_ntPRS_0.001","IQ_ntPRS_0.05","IQ_ntPRS_0.5"),
            ~(scale(.) %>% as.vector))
write.table(scores, "./../output/scores.dat", col.names=T, row.names=F, quote=F, sep ="\t" )
```


Merge with the master Phenotype file

```{r}
master1 = fread("./../output/MasterSet.txt")
master1 = merge (master1, scores, by= "ID", all.x = T)

```
Building Regression Models

```{r}
options ("scipen" = 100, "digits" = 4)

g0 = glm(bin_grade ~ sex + PC1 + PC2 + PC3 + PC4 + PC5 + PC6 + PC7 + PC8 + PC9 + PC10, data= master1 , family = "binomial")

d1 = c("EA_Score_0.001","EA_Score_0.05","EA_Score_0.5","EA_ntPRS_0.001","EA_ntPRS_0.05","EA_ntPRS_0.5",
       "SZ_Score_0.001","SZ_Score_0.05","SZ_Score_0.5","SZ_ntPRS_0.001","SZ_ntPRS_0.05","SZ_ntPRS_0.5",
       "IQ_Score_0.001","IQ_Score_0.05","IQ_Score_0.5","IQ_ntPRS_0.001","IQ_ntPRS_0.05","IQ_ntPRS_0.5")

univars <- data.frame() 

for (i in seq_along(d1)) {
  mod <-
    as.formula(sprintf(
      "bin_grade ~ %s + sex + PC1 + PC2 + PC3 + PC4 + PC5 + PC6 + PC7 + PC8 + PC9 + PC10",
      d1[i]
    ))
  g1 = glm(mod, data = master1 , family = "binomial")
  univars[i, 1] <- names(coef(g1))[2]
  univars[i, 2] <- exp(confint.default(g1)[2, 1])
  univars[i, 3] <- exp(g1$coefficients[2])
  univars[i, 4] <- exp(confint.default(g1)[2, 2])
  univars[i, 5] <- coef(summary(g1))[2, 'Pr(>|z|)']
  univars[i, 6] <- (NagelkerkeR2(g1)$R2 - NagelkerkeR2(g0)$R2) * 100
}

names (univars) = c ("Score" , "Lower_CI", "OR", "Upper_CI", "p-val", "% Variance Explained")

g0 = glm(bin_grade ~ sex, data= master1 , family = "binomial")
d2 = c("EA_pTDT_0.001",
       "EA_pTDT_0.05",
       "EA_pTDT_0.5",
       "SZ_pTDT_0.001",
       "SZ_pTDT_0.05",
       "SZ_pTDT_0.5",
       "IQ_pTDT_0.001",
       "IQ_pTDT_0.05",
       "IQ_pTDT_0.5")

univars2 <- data.frame() 

for (i in seq_along(d2)) {
  mod <- as.formula(sprintf("bin_grade ~ %s + sex", d2[i]))
  g1 = glm(mod, data = master1 , family = "binomial")
  univars2[i, 1] <- names(coef(g1))[2]
  univars2[i, 2] <- exp(confint.default(g1)[2, 1])
  univars2[i, 3] <- exp(g1$coefficients[2])
  univars2[i, 4] <- exp(confint.default(g1)[2, 2])
  univars2[i, 5] <- coef(summary(g1))[2, 'Pr(>|z|)']
  univars2[i, 6] <- (NagelkerkeR2(g1)$R2 - NagelkerkeR2(g0)$R2) * 100
}
names (univars2) = c ("Score" , "Lower_CI", "OR", "Upper_CI", "p-val", "% Variance Explained")
univars = rbind (univars, univars2)
univars
univars = univars %>%
  separate(Score, c("Score", "Type", "Threshold"), "_")
```
Create plot for manuscript

```{r, dpi=200}
univars_t = subset(univars, subset=univars$Threshold==0.05)

univars_t$Type <- factor(univars_t$Type,levels =c("Score","pTDT","ntPRS"))
univars_t = subset (univars_t, subset = univars_t$Type != "pTDT")
plot1 = ggplot(univars_t, aes(Type,OR)) +
geom_errorbar(aes(ymin=Lower_CI, ymax=Upper_CI, color = Score), position = position_dodge(0.4), width = 0.2) +
geom_point(aes(color = Score), position = position_dodge(0.4), size = 2) +
geom_hline(yintercept = 1, linetype = "dashed") +
  scale_color_manual(values = c(wes_palette("Royal1")[1],mypal[3],mypal[4]),
                     labels = c("Educational Attainment","Intelligence", "Schizophrenia")) + 
  scale_x_discrete(labels= c("Proband PRS","Non-Transmitted PRS"))

plot1 + 
  ylab("OR (95% CI)") + 
  xlab("Polygenic scores") +
  theme_classic() +   
  theme(axis.title.y = element_text(face="bold"),
        axis.text.x = element_text(angle = 0, vjust = 0.5, hjust=0.5,face="bold"), 
        axis.title.x = element_text(vjust = -1.3,face="bold"),
        legend.title = element_blank()
        
        )
```

Create Alternate Plot for the Supplement (with parental scores added in)

```{r}
master1 <- master1 %>% 
  mutate_at(c("EA_Mean_Parental_Score_0.001","EA_Mean_Parental_Score_0.05","EA_Mean_Parental_Score_0.5",
              "SZ_Mean_Parental_Score_0.001","SZ_Mean_Parental_Score_0.05","SZ_Mean_Parental_Score_0.5",
              "IQ_Mean_Parental_Score_0.001","IQ_Mean_Parental_Score_0.05","IQ_Mean_Parental_Score_0.5"), 
            ~(scale(.) %>% as.vector))


d4 = c("EA_Mean_Parental_Score_0.001","EA_Mean_Parental_Score_0.05","EA_Mean_Parental_Score_0.5",
       "SZ_Mean_Parental_Score_0.001","SZ_Mean_Parental_Score_0.05","SZ_Mean_Parental_Score_0.5",
       "IQ_Mean_Parental_Score_0.001","IQ_Mean_Parental_Score_0.05","IQ_Mean_Parental_Score_0.5")

g0 = glm(bin_grade ~ sex + PC1 + PC2 + PC3 + PC4 + PC5 + PC6 + PC7 + PC8 + PC9 + PC10, data= master1 , family = "binomial")

univars3 <- data.frame() 

for (i in seq_along(d4)) {
  mod <-
    as.formula(sprintf(
      "bin_grade ~ %s + sex + PC1 + PC2 + PC3 + PC4 + PC5 + PC6 + PC7 + PC8 + PC9 + PC10",
      d4[i]
    ))
  g1 = glm(mod, data = master1 , family = "binomial")
  univars3[i, 1] <- names(coef(g1))[2]
  univars3[i, 2] <- exp(confint.default(g1)[2, 1])
  univars3[i, 3] <- exp(g1$coefficients[2])
  univars3[i, 4] <- exp(confint.default(g1)[2, 2])
  univars3[i, 5] <- coef(summary(g1))[2, 'Pr(>|z|)']
  univars3[i, 6] <- (NagelkerkeR2(g1)$R2 - NagelkerkeR2(g0)$R2) * 100
}

names (univars3) = c ("Score" , "Lower_CI", "OR", "Upper_CI", "p-val", "% Variance Explained")

univars3 = univars3 %>%
  separate(Score, c("Score","m1" ,"Type","m2", "Threshold"), "_")

univars3 = univars3[,-c(2,4)]
univars3
univars= rbind (univars,univars3)
```
```{r, dpi=200}
univars_t2 = subset(univars, subset=univars$Threshold==0.05)

univars_t2$Type <- factor(univars_t2$Type,levels =c("Score","Parental","pTDT","ntPRS"))


plot2 = ggplot(univars_t2, aes(Score,OR)) +
geom_errorbar(aes(ymin=Lower_CI, ymax=Upper_CI, color = Type), position = position_dodge(0.45), width = 0.2)+
geom_point(aes(color = Type), position = position_dodge(0.45), size = 2) +
geom_hline(yintercept = 1, linetype = "dashed") +
  scale_color_manual(values = c(wes_palette("Royal1")[1],mypal[3],mypal[4],wes_palette("Royal1")[2]),
                     labels = c("Proband PRS","Parental-PRS","pTDT","nt-PRS")) 

plot2 + 
  ylab("OR (95% CI)") + 
  xlab("Polygenic scores") +
  theme_classic() +   
  theme(axis.title.y = element_text(face="bold"),
        axis.text.x = element_text(angle = 45, vjust = 0.5, hjust=0.5,face="bold"), 
        axis.title.x = element_text(vjust = -1.3,face="bold"),
        legend.title = element_blank()
        )

```
Does it change if parents are examined separately?
```{r}

pca1 = read.table("./../data/Trios_PCA1.eigenvec", head=F)
names(pca1)[3:ncol(pca1)] = paste0("PC", 1:(ncol(pca1)-2))
names (pca1)[1] = "Family"
names (pca1)[2] = "ID"
pca1 = transform(pca1,Family_Frequency=ave(seq(nrow(pca1)),Family,FUN=length))
pca1 = subset (pca1, subset= pca1$Family_Frequency>2)
pca1 = pca1 %>%separate(ID, c("FID", "mem"),"-", remove = FALSE)
pca1 = subset(pca1, subset=pca1$mem==2|pca1$mem==3 )

pc1 = read.table ("./../data/EA_Trios.all.score", head=T)
pc2 = read.table ("./../data/SZ_Trios.all.score", head=T)
pc3 = read.table ("./../data/IQ_Trios.all.score", head=T)

pc1_2 = subset(pc1, subset=pc1$MEMBERSHIP == "-2"|pc1$MEMBERSHIP == "-3")

pc2_2 = subset(pc2, subset=pc2$MEMBERSHIP == "-2"|pc2$MEMBERSHIP == "-3")
pc3_2 = subset(pc3, subset=pc3$MEMBERSHIP == "-2"|pc3$MEMBERSHIP == "-3")

names(pc1_2) = c("FID","ID","MEMBERSHIP", "EA_Score_0.001", "EA_Score_0.05",  "EA_Score_0.5", "EA_Score_1")
names(pc2_2) = c("FID","ID","MEMBERSHIP", "SZ_Score_0.001", "SZ_Score_0.05",  "SZ_Score_0.5", "SZ_Score_1")
names(pc3_2) = c("FID","ID","MEMBERSHIP", "IQ_Score_0.001", "IQ_Score_0.05",  "IQ_Score_0.5", "IQ_Score_1")

par1 = merge (pc1_2, pc2_2, by = "ID")
par1 = merge (par1, pc3_2, by = "ID")
par1 = merge (par1, pca1, by = "ID")
par1 = par1[,-c(2,8,14,21)]
par1$Family = as.character(par1$Family)
master2 = fread("./../output/MasterSet.txt")
master2$Family.x = as.character(master2$Family.x)
master2 = merge (master2, par1, by.x= "Family.x", by.y="Family")
master2 <- master2 %>% 
  mutate_at(c("EA_Score_0.001","EA_Score_0.05","EA_Score_0.5",
              "SZ_Score_0.001","SZ_Score_0.05","SZ_Score_0.5",
              "IQ_Score_0.001","IQ_Score_0.05","IQ_Score_0.5"), 
            ~(scale(.) %>% as.vector))


d5 = c("EA_Score_0.001","EA_Score_0.05","EA_Score_0.5",
              "SZ_Score_0.001","SZ_Score_0.05","SZ_Score_0.5",
              "IQ_Score_0.001","IQ_Score_0.05","IQ_Score_0.5")

master2= master2%>% mutate (MEMBERSHIP2 = recode(MEMBERSHIP,
"-2" = 0,
"-3" = 1,
))

g0 = glm(bin_grade ~ MEMBERSHIP2 + PC1.y + PC2.y + PC3.y + PC4.y + PC5.y + PC6.y + PC7.y + PC8.y + PC9.y + PC10.y, data= master2 , family = "binomial")

univars4 <- data.frame() 

for (i in seq_along(d5)) {
  mod <-
    as.formula(sprintf(
      "bin_grade ~ %s + MEMBERSHIP2 + PC1.y + PC2.y + PC3.y + PC4.y + PC5.y + PC6.y + PC7.y + PC8.y + PC9.y + PC10.y",
      d5[i]
    ))
  g1 = glm(mod, data = master2 , family = "binomial")
  univars4[i, 1] <- names(coef(g1))[2]
  univars4[i, 2] <- exp(confint.default(g1)[2, 1])
  univars4[i, 3] <- exp(g1$coefficients[2])
  univars4[i, 4] <- exp(confint.default(g1)[2, 2])
  univars4[i, 5] <- coef(summary(g1))[2, 'Pr(>|z|)']
  univars4[i, 6] <- (NagelkerkeR2(g1)$R2 - NagelkerkeR2(g0)$R2) * 100
}

names (univars4) = c ("Parent_Score" , "Lower_CI", "OR", "Upper_CI", "p-val", "% Variance Explained")

univars4

```
